// apps/matching/src/types/matching.ts
export interface RecruitmentBase {
  recruitmentId: number;
  title: string;
  workDays: WorkDay[];
  workStartTime: string;
  workEndTime: string;
  workSalaryUnitType: WorkSalaryUnitType;
  workSalaryAmount: number;
}

export interface RecruitmentForm
  extends Omit<RecruitmentBase, 'recruitmentId'> {
  elderlyId: number | null;
  careTypes: string[];
  description: string;
}